Fix table range edges. Add DEBUG constants.
Fixed an issue with the table range edges, caused by not having
valid data above or to the left of the table range. In the
square problem, we fill with zeros along these edges, and now
the code fills zeros along the limiting left and upper edge
efficiently. Wasn't an issue with negative sentinels in the
table, but once those were removed for speed, problems came up
quickly.

Also add DEBUG constants to turn on and off certain features, such
as sentinels and printing features. These should compile completely
out when false.

- Judson