Fixes: Bottom corner start, and Aext vs A
The first fix corrects an issue where the backward path traversal was
not starting in the lower right most corner of the valid table
entries every time, becaues there are 2 different boundary conditions,
the m-wide band boundary, and the lower path limit boundary.

The second fix was needed when comparing A and B chars in the lower
half of the table, indexing into A instead of Aext.

- Judson