Complete, but broken, implementation.
Present issue appears to be that when the upper and lower boundaries
are the same, there is no room for the table to propogate, so I need
to ensure the U,D,L,R constraints of a path allow for unit width
when the path is horizontal or vertical (diagonal is OK).

Fails on sample.in test #2.